+++ internal/model/default_values.go
@@ -21,6 +21,7 @@ var NodeSelectorArgs, ImagePullSecretArgs, CustomApocImageArgs, PriorityClassNam
 
 var NodeSelectorLabel = "testLabel=1"
 var DefaultNeo4jName = "test-cluster"
+var DefaultNeo4jChartName = "neo4j"
 var DefaultClusterSize = 3
 var DefaultNeo4jNameArg = []string{"--set", "neo4j.name=" + DefaultNeo4jName}
 var DefaultClusterSizeArg = []string{"--set", "neo4j.minimumClusterSize=" + strconv.Itoa(DefaultClusterSize)}
